added routes.ts
