+++ src/index.ts
@@ -1,7 +1,8 @@
-import express, { Application, Request, Response } from 'express';
+import express, { Application } from 'express';
 import dotenv from 'dotenv';
 import fs from 'fs';
 import path from 'path';
+import { setupRoutes } from './routes';
 
 // Load environment variables
 dotenv.config();
@@ -21,52 +22,6 @@ const loadMappings = (): { [key: string]: number } => {
   }
 };
 
-// Ensure data directory exists
-const dataDir = path.join(__dirname, '..', 'data');
-if (!fs.existsSync(dataDir)) {
-  fs.mkdirSync(dataDir, { recursive: true });
-}
-
-// Logging utility
-const logRequestResponse = (prompt: any, response: any, endpoint: string, workflowId?: string) => {
-  // Save to JSONL file in ChatML format
-  try {
-    const chatMLEntry = {
-      messages: [
-        {
-          role: "user",
-          content: prompt.prompt || JSON.stringify(prompt)
-        },
-        {
-          role: "assistant", 
-          content: response.choices?.[0]?.text || response.choices?.[0]?.message?.content || JSON.stringify(response)
-        }
-      ],
-      timestamp: new Date().toISOString(),
-      endpoint: endpoint,
-      model: prompt.model || 'unknown',
-      workflow_id: workflowId || 'default'
-    };
-
-    // Create workflow-specific directory
-    const workflowDir = workflowId 
-      ? path.join(dataDir, workflowId)
-      : path.join(dataDir, 'default');
-    
-    if (!fs.existsSync(workflowDir)) {
-      fs.mkdirSync(workflowDir, { recursive: true });
-    }
-
-    const jsonlLine = JSON.stringify(chatMLEntry) + '\n';
-    const filePath = path.join(workflowDir, 'data.jsonl');
-    
-    fs.appendFileSync(filePath, jsonlLine);
-    console.log(`Saved conversation to ${filePath} (workflow: ${workflowId || 'default'})`);
-  } catch (error) {
-    console.error('Error saving conversation:', error);
-  }
-};
-
 // Create an Express app with workflow-specific routing
 const createApp = (workflowId: string): Application => {
   const app: Application = express();
@@ -75,102 +30,8 @@ const createApp = (workflowId: string): Application => {
   app.use(express.json({ limit: '10mb' }));
   app.use(express.urlencoded({ extended: true }));
 
-  // Proxy route for OpenAI /v1/completions
-  app.post('/v1/completions', async (req: Request, res: Response) => {
-    try {
-      const requestBody = req.body;
-      // Use the workflow_id from the port mapping, but allow override from request body
-      const finalWorkflowId = requestBody.workflow_id || workflowId;
-      
-      // Remove workflow_id from request body before forwarding to API
-      const forwardBody = { ...requestBody };
-      delete forwardBody.workflow_id;
-      
-      // Log the incoming request
-      console.log(`Proxying request to ${OPENAI_API_BASE_URL}/v1/completions for ${finalWorkflowId}`);
-      
-      // Forward request to OpenAI API
-      const controller = new AbortController();
-      const timeoutId = setTimeout(() => controller.abort(), 60000); // 1 minute timeout
-      
-      // Prepare headers, filtering out problematic ones
-      const forwardHeaders: Record<string, string> = {};
-      Object.entries(req.headers).forEach(([key, value]) => {
-        if (typeof value === 'string' && key.toLowerCase() !== 'host' && key.toLowerCase() !== 'content-length') {
-          forwardHeaders[key] = value;
-        }
-      });
-      
-      const response = await fetch(
-        `${OPENAI_API_BASE_URL}/v1/completions`,
-        {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-            ...forwardHeaders
-          },
-          body: JSON.stringify(forwardBody),
-          signal: controller.signal
-        }
-      );
-      
-      clearTimeout(timeoutId);
-      
-      if (!response.ok) {
-        const errorData = await response.json();
-        return res.status(response.status).json(errorData);
-      }
-      
-      const responseData = await response.json();
-
-      // Log the prompt and response with workflow_id
-      logRequestResponse(requestBody, responseData, '/v1/completions', finalWorkflowId);
-
-      // Return the response to the client
-      res.status(response.status).json(responseData);
-
-    } catch (error: any) {
-      console.error('Error proxying request:', error.message);
-      
-      if (error.name === 'AbortError') {
-        // Request was aborted due to timeout
-        res.status(408).json({ 
-          error: 'Request timeout',
-          message: 'Request timed out after 60 seconds'
-        });
-      } else {
-        // Internal server error
-        res.status(500).json({ 
-          error: 'Internal proxy server error',
-          message: error.message 
-        });
-      }
-    }
-  });
-
-  // Health check endpoint
-  app.get('/health', (req: Request, res: Response) => {
-    res.json({ 
-      status: 'OK', 
-      timestamp: new Date().toISOString(),
-      service: 'OpenAI Proxy Service',
-      workflow_id: workflowId
-    });
-  });
-
-  // Basic info endpoint
-  app.get('/', (req: Request, res: Response) => {
-    res.json({
-      service: 'OpenAI Proxy Service',
-      description: 'Forwards requests to OpenAI API while logging prompts and responses',
-      workflow_id: workflowId,
-      endpoints: {
-        'POST /v1/completions': 'Proxy for OpenAI chat completions',
-        'GET /health': 'Health check',
-        'GET /': 'Service information'
-      }
-    });
-  });
+  // Setup routes
+  setupRoutes(app, workflowId);
 
   return app;
 };
